commands: improve value stat commands
